tests/wpt/web-platform-tests/html/canvas/element/line-styles/2d.line.cap.butt.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>Canvas test: 2d.line.cap.butt</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.line.cap.butt</h1>
<p class="desc">lineCap 'butt' is rendered correctly</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:<p><img src="/images/green-100x50.png" class="output expected" id="expected" alt="">
<ul id="d"></ul>
<script>
var t = async_test("lineCap 'butt' is rendered correctly");
_addTest(function(canvas, ctx) {

ctx.fillStyle = '#0f0';
ctx.fillRect(0, 0, 100, 50);

ctx.lineCap = 'butt';
ctx.lineWidth = 20;

ctx.fillStyle = '#f00';
ctx.strokeStyle = '#0f0';
ctx.fillRect(15, 15, 20, 20);
ctx.beginPath();
ctx.moveTo(25, 15);
ctx.lineTo(25, 35);
ctx.stroke();

ctx.fillStyle = '#0f0';
ctx.strokeStyle = '#f00';
ctx.beginPath();
ctx.moveTo(75, 15);
ctx.lineTo(75, 35);
ctx.stroke();
ctx.fillRect(65, 15, 20, 20);

_assertPixel(canvas, 25,14, 0,255,0,255, "25,14", "0,255,0,255");
_assertPixel(canvas, 25,15, 0,255,0,255, "25,15", "0,255,0,255");
_assertPixel(canvas, 25,16, 0,255,0,255, "25,16", "0,255,0,255");
_assertPixel(canvas, 25,34, 0,255,0,255, "25,34", "0,255,0,255");
_assertPixel(canvas, 25,35, 0,255,0,255, "25,35", "0,255,0,255");
_assertPixel(canvas, 25,36, 0,255,0,255, "25,36", "0,255,0,255");

_assertPixel(canvas, 75,14, 0,255,0,255, "75,14", "0,255,0,255");
_assertPixel(canvas, 75,15, 0,255,0,255, "75,15", "0,255,0,255");
_assertPixel(canvas, 75,16, 0,255,0,255, "75,16", "0,255,0,255");
_assertPixel(canvas, 75,34, 0,255,0,255, "75,34", "0,255,0,255");
_assertPixel(canvas, 75,35, 0,255,0,255, "75,35", "0,255,0,255");
_assertPixel(canvas, 75,36, 0,255,0,255, "75,36", "0,255,0,255");


});
</script>
